:bug: fix selectedModelForStudy selection problem
